Pass in core.Application as BasePluginTest's application context
Because the abstract `getTestGrailsApplicationClass()` is called in a static context, the method is treated as `static` which does now allow itself to be overridden.

Hiding it by defining the same static method in the `PostgresImporterTest` subclass does not work either because the compiler determines the (empty) parent method to be run first. Hence the resulting `MissingMethodException`.

This change is a temporary fix in order to get plugin tests running first. It still needs further investigation.